Create Transaction Provider
Context engine for handling transactions. Each transaction takes several seconds, so decent handling is needed.

This change also cleans up Profiles, pulling ProfileForm into its own component and cleaning up warnings.
