--- src/components/SetProfile.js
@@ -1,9 +1,0 @@
-
-
-const updateProfile = (address, newProfile) => {
-    if (!address || !newProfile ) throw new Error('Null address/profile passed to Profile set function')
-
-    //TODO trigger Flow upload
-}
-
-export default updateProfile
